requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain_huggingface
langchain_chroma
langchain-groq
python-dotenv
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
chromadb
sentence_transformers
faiss-cpu
duckdb
pandas
duckduckgo-search
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
psycopg2-binary
sqlalchemy